packages/datacommons-mcp/proof_of_concept/requirements.txt
fastmcp
asyncio
datacommons-client
requests
python-dotenv>=1.0.0
pydantic
fastapi
uvicorn
uv